flag d et dump tous les cycles fini

+++ VM/includes/vm.h
@@ -6,7 +6,7 @@
 /*   By: toliver <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/12/04 01:38:07 by toliver           #+#    #+#             */
-/*   Updated: 2019/12/16 06:59:53 by toliver          ###   ########.fr       */
+/*   Updated: 2019/12/19 03:41:33 by toliver          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -139,7 +139,7 @@ typedef struct		s_arena
 	t_process		*process;
 	WINDOW			*main;
 	WINDOW			*infos;
-	long long int	cycles;
+	size_t			cycles;
 	int				mode;
 	int				cycles_to_die;
 }					t_arena;
